+++ packages/ricos-types/web/src/index.ts
@@ -24,4 +24,3 @@ export * from './shortcuts';
 export * from './key-types';
 export * from './RicosTypes';
 export * from './events';
-export * from './contentResolverTypes';