+++ packages/toolbars-v3/web/src/modals/FloatingPluginMenu/AddPluginMenuWrapper.tsx
@@ -60,7 +60,6 @@ const AddPluginMenuWrapper: React.FC<Props> = ({
 
   const onPluginMenuButtonClick: IPluginMenuButtonClick = (modal, command) => {
     modalService.closeModal(PLUGIN_MENU_MODAL_ID);
-    modalService.openModal;
     return modal
       ? modalService?.openModal({
           Component: modal.Component,